:sparkles: Feat(profile-menu): 추갸 유저 아바타 클릭 시 메뉴 팝업
Related to: #104
