+++ src/components/header.tsx
@@ -1,6 +1,6 @@
 import { Logo } from '@/components/logo'
 import { Navigation } from '@/components/navigation'
-import { ProfileMenu } from '@/components/profile-menu'
+import { ProfileMenu } from '@/components/user/profile-menu'
 import { MobileNavigation } from '@/components/mobile-navigation'
 
 export const Header = () => {
